requirements.txt
alembic==1.13.0
amqp==5.1.1
async-timeout==4.0.3
autopep8==2.0.4
billiard==4.1.0
blinker==1.6.3
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.5
flake8==6.1.0
Flask==3.0.0
Flask-JWT-Extended==4.5.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.2
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mccabe==0.7.0
packaging==23.2
passlib==1.7.4
prompt-toolkit==3.0.39
psutil==5.9.6
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
setuptools==65.5.0
six==1.16.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.8
Werkzeug==3.0.0
WTForms==3.1.1
